interview-stuff/S2/pyproject.toml
[project]
name = "s2"
version = "0.1.0"
description = "Deep agents FastAPI harness"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.4.0",
]
